+++ lib/src/rtc_session.dart
@@ -130,10 +130,10 @@ class RTCSession extends EventManager implements Owner {
   String? _tones;
 
   // Mute/Hold state.
-  bool? _audioMuted;
-  bool? _videoMuted;
-  bool? _localHold;
-  bool? _remoteHold;
+  bool _audioMuted = false;
+  bool _videoMuted = false;
+  bool _localHold = false;
+  bool _remoteHold = false;
 
   late RFC4028Timers _sessionTimers;
 
@@ -956,23 +956,22 @@ class RTCSession extends EventManager implements Owner {
    */
   void mute([bool audio = true, bool video = true]) {
     logger.d('mute()');
+    bool changed = false;
 
-    bool audioMuted = false, videoMuted = false;
-
-    if (_audioMuted == false && audio) {
-      audioMuted = true;
+    if (!_audioMuted && audio) {
       _audioMuted = true;
+      changed = true;
       _toggleMuteAudio(true);
     }
 
-    if (_videoMuted == false && video) {
-      videoMuted = true;
+    if (!_videoMuted && video) {
       _videoMuted = true;
+      changed = true;
       _toggleMuteVideo(true);
     }
 
-    if (audioMuted == true || videoMuted == true) {
-      _onmute(audioMuted, videoMuted);
+    if (changed) {
+      _onmute(_audioMuted, _videoMuted);
     }
   }
 
@@ -981,29 +980,28 @@ class RTCSession extends EventManager implements Owner {
    */
   void unmute([bool audio = true, bool video = true]) {
     logger.d('unmute()');
-
-    bool audioUnMuted = false, videoUnMuted = false;
+    bool changed = false;
 
     if (_audioMuted == true && audio) {
-      audioUnMuted = true;
       _audioMuted = false;
 
       if (_localHold == false) {
+        changed = true;
         _toggleMuteAudio(false);
       }
     }
 
     if (_videoMuted == true && video) {
-      videoUnMuted = true;
       _videoMuted = false;
 
       if (_localHold == false) {
+        changed = true;
         _toggleMuteVideo(false);
       }
     }
 
-    if (audioUnMuted == true || videoUnMuted == true) {
-      _onunmute(audioUnMuted, videoUnMuted);
+    if (changed) {
+      _onunmute(!_audioMuted, !_videoMuted);
     }
   }
 
@@ -1554,8 +1552,11 @@ class RTCSession extends EventManager implements Owner {
   }
 
   void _iceRestart() async {
-    Map<String, dynamic> offerConstraints =
-        _rtcOfferConstraints ?? <String, dynamic>{};
+    Map<String, dynamic> offerConstraints = _rtcOfferConstraints ??
+        <String, dynamic>{
+          'mandatory': <String, dynamic>{},
+          'optional': <dynamic>[],
+        };
     offerConstraints['mandatory']['IceRestart'] = true;
     renegotiate(offerConstraints);
   }
@@ -1612,6 +1613,18 @@ class RTCSession extends EventManager implements Owner {
     Completer<RTCSessionDescription> completer =
         Completer<RTCSessionDescription>();
 
+    constraints = constraints ??
+        <String, dynamic>{
+          'mandatory': <String, dynamic>{},
+          'optional': <dynamic>[],
+        };
+
+    List<Future<RTCSessionDescription> Function(RTCSessionDescription)>
+        modifiers = constraints['offerModifiers'] ??
+            <Future<RTCSessionDescription> Function(RTCSessionDescription)>[];
+
+    constraints['offerModifiers'] = null;
+
     if (type != 'offer' && type != 'answer') {
       completer.completeError(Exceptions.TypeError(
           'createLocalDescription() | invalid type "$type"'));
@@ -1621,7 +1634,7 @@ class RTCSession extends EventManager implements Owner {
     late RTCSessionDescription desc;
     if (type == 'offer') {
       try {
-        desc = await _connection!.createOffer(constraints!);
+        desc = await _connection!.createOffer(constraints);
       } catch (error) {
         logger.e(
             'emit "peerconnection:createofferfailed" [error:${error.toString()}]');
@@ -1630,7 +1643,7 @@ class RTCSession extends EventManager implements Owner {
       }
     } else {
       try {
-        desc = await _connection!.createAnswer(constraints!);
+        desc = await _connection!.createAnswer(constraints);
       } catch (error) {
         logger.e(
             'emit "peerconnection:createanswerfailed" [error:${error.toString()}]');
@@ -1641,6 +1654,11 @@ class RTCSession extends EventManager implements Owner {
 
     // Add 'pc.onicencandidate' event handler to resolve on last candidate.
     bool finished = false;
+
+    for (Future<RTCSessionDescription> Function(RTCSessionDescription) modifier
+        in modifiers) {
+      desc = await modifier(desc);
+    }
 
     Future<void> ready() async {
       if (!finished && _status != C.STATUS_TERMINATED) {
@@ -1675,7 +1693,9 @@ class RTCSession extends EventManager implements Owner {
            *  Because trickle ICE is not defined in the sip protocol, the delay of
            * initiating a call to answer the call waiting will be unacceptable.
            */
-          setTimeout(() => ready(), ua.configuration.ice_gathering_timeout);
+          if (ua.configuration.ice_gathering_timeout != 0) {
+            setTimeout(() => ready(), ua.configuration.ice_gathering_timeout);
+          }
         }
       }
     };
@@ -2356,7 +2376,7 @@ class RTCSession extends EventManager implements Owner {
           RTCSessionDescription(response.body, 'answer');
 
       try {
-        _connection!.setRemoteDescription(answer);
+        await _connection!.setRemoteDescription(answer);
       } catch (error) {
         logger.e(
             'emit "peerconnection:setremotedescriptionfailed" [error:${error.toString()}]');
@@ -2703,14 +2723,14 @@ class RTCSession extends EventManager implements Owner {
    * Correctly set the SDP direction attributes if the call is on local hold
    */
   String? _mangleOffer(String? sdpInput) {
-    if (!_localHold! && !_remoteHold!) {
+    if (!_localHold && !_remoteHold) {
       return sdpInput;
     }
 
     Map<String, dynamic> sdp = sdp_transform.parse(sdpInput!);
 
     // Local hold.
-    if (_localHold! && !_remoteHold!) {
+    if (_localHold && !_remoteHold) {
       logger.d('mangleOffer() | me on hold, mangling offer');
       for (Map<String, dynamic> m in sdp['media']) {
         if (holdMediaTypes.indexOf(m['type']) == -1) {
@@ -2726,7 +2746,7 @@ class RTCSession extends EventManager implements Owner {
       }
     }
     // Local and remote hold.
-    else if (_localHold! && _remoteHold!) {
+    else if (_localHold && _remoteHold) {
       logger.d('mangleOffer() | both on hold, mangling offer');
       for (Map<String, dynamic> m in sdp['media']) {
         if (holdMediaTypes.indexOf(m['type']) == -1) {
@@ -2736,7 +2756,7 @@ class RTCSession extends EventManager implements Owner {
       }
     }
     // Remote hold.
-    else if (_remoteHold!) {
+    else if (_remoteHold) {
       logger.d('mangleOffer() | remote on hold, mangling offer');
       for (Map<String, dynamic> m in sdp['media']) {
         if (holdMediaTypes.indexOf(m['type']) == -1) {
@@ -2758,16 +2778,16 @@ class RTCSession extends EventManager implements Owner {
   void _setLocalMediaStatus() {
     bool enableAudio = true, enableVideo = true;
 
-    if (_localHold! || _remoteHold!) {
+    if (_localHold || _remoteHold) {
       enableAudio = false;
       enableVideo = false;
     }
 
-    if (_audioMuted!) {
+    if (_audioMuted) {
       enableAudio = false;
     }
 
-    if (_videoMuted!) {
+    if (_videoMuted) {
       enableVideo = false;
     }