+++ CHANGELOG.md
@@ -1,6 +1,25 @@
 # Changelog
 
 --------------------------------------------
+[0.5.5] - 2023.03.08
+
+* Bump version for intl & lints
+* Update websocket_web_impl.dart (#345)
+* fix(hangup): set cancel reason nullable (#346)
+* Add sdp transformers (#350)
+* Hold fix (#351)
+
+[0.5.4] - 2023.02.20
+
+* Bump version for flutter-webrtc
+* Fixed error handling in _receiveInviteResponse #344
+* Updated logger #342
+* Websocket message queue using streams and Delay between messages #335
+* Fixed bugs in message.dart & rtc_session.dart #332
+* Allow setting ice_gathering_timeout option #330
+* Add realm option to SIPUAHelper settings #331
+* Code quality #326
+
 [0.5.3] - 2022.10.19
 
 * Bump version for flutter-webrtc